pyproject.toml
[project]
name = "investidor-ia"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.5.0",
    "anthropic>=0.49.0",
    "beautifulsoup4>=4.13.3",
    "diskcache>=5.6.3",
    "dotenv>=0.9.9",
    "duckduckgo-search>=7.5.1",
    "fpdf2>=2.8.4",
    "google-genai>=1.5.0",
    "groq>=1.1.2",
    "lancedb>=0.30.2",
    "markdown>=3.7",
    "markdown2>=2.5.3",
    "mcp>=1.27.0",
    "nest-asyncio>=1.6.0",
    "openai>=1.69.0",
    "pandas>=2.2.3",
    "polars>=1.24.0",
    "pyarrow>=19.0.1",
    "pydantic>=2.10.6",
    "pymupdf>=1.25.3",
    "pypdf>=5.5.0",
    "python-decouple>=3.8",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.43.2",
    "unidecode>=1.3.8",
    "yfinance>=0.2.54",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.0.0",
    "ruff>=0.15.9",
]
